requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
botocore==1.36.12
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==4.0.0
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
Cython==3.0.12
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
Django==5.1.7
docstring_parser==0.16
executing==2.2.0
faiss-cpu==1.10.0
filelock==3.17.0
firebase-admin==6.7.0
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.1.24
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.86.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
ibm-cos-sdk==2.14.0
ibm-cos-sdk-core==2.14.0
ibm-cos-sdk-s3transfer==2.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
iso8601==2.1.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-text-splitters==0.3.6
langsmith==0.3.13
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
openai==1.66.3
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyDirectInput==1.0.4
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.3.1
pypdfium2==4.30.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==4.0.1
serial==0.0.97
setuptools==75.8.0
shapely==2.0.7
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.3
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
